simplify imports from builtin actors
